refactor: share opportunity summary aggregation
