+++ backend/app/infra/db/uow.py
@@ -14,6 +14,9 @@ from sqlalchemy.orm import Session, sessionmaker
 from app.domain.common.uow import UnitOfWork
 from app.infra.db.repositories.feature_run_repo import SqlFeatureRunRepository
 from app.infra.db.repositories.feature_store_repo import SqlFeatureStoreRepository
+from app.infra.db.repositories.opportunity_summary_repo import (
+    SqlOpportunityStateSummaryRepository,
+)
 from app.infra.db.repositories.scan_repo import SqlScanRepository
 from app.infra.db.repositories.scan_result_repo import SqlScanResultRepository
 from app.infra.db.repositories.universe_repo import SqlUniverseRepository
@@ -29,6 +32,7 @@ class SqlUnitOfWork(UnitOfWork):
         self.session: Session = self._session_factory()
         self.scans = SqlScanRepository(self.session)
         self.scan_results = SqlScanResultRepository(self.session)
+        self.opportunity_summaries = SqlOpportunityStateSummaryRepository(self.session)
         self.universe = SqlUniverseRepository(self.session)
         self.feature_runs = SqlFeatureRunRepository(self.session)
         self.feature_store = SqlFeatureStoreRepository(self.session)
